refactor: question.tsxの切り分けられる部分コンポーネントに分離
